Adjust auto-update to trigger on TF pre-releases
This commit makes some changes to the auto-update
and release mechanism. Firstly, we move to 'yq'
for grabbing the version of the snap in various places
and reduce our reliance on regex.

Secondly, we now also allow the auto updated to run
on pre-release versions of Terraform. Accordingly, the
automatic release logic has been updated such that the
pre-release versions are just released to edge, and not
to candidate.